% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_dir_slash}
\alias{append_dir_slash}
\title{Adds a slash at the end of path}
\usage{
append_dir_slash(dir_string)
}
\arguments{
\item{dir_string}{character string representing path to directory}
}
\value{
path too directory with slash as a last character
}
\description{
Slash is added only if there is none present.
}
