% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{collate_cp_pathways}
\alias{collate_cp_pathways}
\title{Batch read files from  clusterprofiler_reports_snakemake into one data frame}
\usage{
collate_cp_pathways(
  pathway_files_basepath,
  pattern,
  pval_threshold = 0.05,
  recursive = TRUE
)
}
\arguments{
\item{pathway_files_basepath}{character string, directory path where the
result files are located.}

\item{pattern}{character string, regex pattern for filenames to include.}

\item{padj_threshold}{double, p-value significance threshold fo filtering
pathways}
}
\value{
data frame
}
\description{
Batch read files from  clusterprofiler_reports_snakemake into one data frame
}
