% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{collate_dire_pathways}
\alias{collate_dire_pathways}
\title{Batch reading of dire data from directory}
\usage{
collate_dire_pathways(
  pathway_files_basepath,
  pattern = "",
  sheet_name = "Sheet1",
  recursive = TRUE
)
}
\arguments{
\item{pathway_files_basepath}{character string, directory where dire data is
located.}

\item{pattern}{character string, regular expression to match the filenames
in specified directory, empty string will match all files.
default: empty string}

\item{sheet_name}{character string, for excel files the name of the sheet
that contains the dire data. default: "Sheet1"}
}
\value{
data frame
}
\description{
Function will attempt to read all text and excel files in directory that
match the specified pattern. If other types of files match, it will throw
an error.
}
