% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{collate_ipa_pathways}
\alias{collate_ipa_pathways}
\title{Batch read IPA files into one data frame}
\usage{
collate_ipa_pathways(
  pathway_files_basepath,
  pattern,
  rank_by = zscore,
  descending = TRUE,
  recursive = TRUE
)
}
\arguments{
\item{pathway_files_basepath}{character string, directory path where the
result files are located.}

\item{pattern}{character string, regex pattern for filenames to include.}

\item{rank_by}{which column to rank the data on, supplied as variable.
default: zscore}

\item{descending}{logical, whether ranking should be done from highest to
lowest value. default: TRUE}
}
\value{
data frame with IPA pathways
}
\description{
Batch read IPA files into one data frame
}
