% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{collate_pathways_amigo}
\alias{collate_pathways_amigo}
\title{Batch reading of AmiGO data from directory'}
\usage{
collate_pathways_amigo(
  dir,
  pattern = "",
  rank_by = pvalue,
  filter_level_up_to = 0,
  recursive = TRUE
)
}
\arguments{
\item{dir}{character string, directory path where the
result files are located.}

\item{pattern}{character string, regex pattern for filenames to include.}

\item{rank_by}{which column to rank the data on, supplied as variable.
default: pvalue}

\item{filter_level_up_to}{integer, filter hierarchical levels for pathways.
Root hierarchy level is numbered with 0. Will filter levels up to the
specified hierarchy level, if NULL no filtering will be performed.
default: 0}

\item{recursive}{logical, whether directories should be scanned recursively
for pathway files. default: TRUE}
}
\value{
data frame
}
\description{
Batch reading of AmiGO data from directory'
}
