% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{collate_pathways_cp}
\alias{collate_pathways_cp}
\title{Batch read files from  clusterprofiler_reports_snakemake into one data frame}
\usage{
collate_pathways_cp(dir, pattern = "", pval_threshold = 0.05, recursive = TRUE)
}
\arguments{
\item{dir}{character string, directory path where the
result files are located.}

\item{pattern}{character string, regex pattern for filenames to include.}

\item{pval_threshold}{double, p-value significance threshold fo filtering
pathways}

\item{recursive}{logical, whether directories should be scanned recursively
for pathway files. default: TRUE}
}
\value{
data frame
}
\description{
Batch read files from  clusterprofiler_reports_snakemake into one data frame
}
