% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{collate_pathways_gsea}
\alias{collate_pathways_gsea}
\title{Batch read GSEA files into one data frame}
\usage{
collate_pathways_gsea(
  dir,
  pattern = "gsea_report_for",
  score_column = nes,
  score_threshold = 0,
  pvalue_column = fdr_qval,
  pvalue_threshold = 0.05,
  rank_by = nes,
  recursive = TRUE
)
}
\arguments{
\item{dir}{character string, directory with GSEA results to process. Will be
processed recursively. default: . (current directory)}

\item{pattern}{character string, regex pattern for filenames to include.
default: gsea_report_for}

\item{score_column}{name of score column, supplied as variable. default: nes}

\item{score_threshold}{double, return data subset that pass the score
threshold. Will compare the threshold against absolute values from score
column. default: 0}

\item{pvalue_column}{name of p-value column, supplied as variable. default:
fdr_qval}

\item{pvalue_threshold}{double, return data subset that pass the p-value
threshold.}

\item{rank_by}{which column to rank the data on, supplied as variable.
default: nes}

\item{recursive}{logical, whether directories should be scanned recursively
for pathway files. default: TRUE}
}
\value{
data frame with parsed GSEA data
}
\description{
Combines get_gsea_files_from_dir, read_pathways_gsea and filter_pathways
functions and sets the column and threshold defaults.
}
