% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{collate_pathways_ipa}
\alias{collate_pathways_ipa}
\title{Batch read IPA files into one data frame}
\usage{
collate_pathways_ipa(
  dir,
  pattern,
  rank_by = zscore,
  descending = TRUE,
  recursive = TRUE
)
}
\arguments{
\item{dir}{character string, directory path where the
result files are located.}

\item{pattern}{character string, regex pattern for filenames to include.}

\item{rank_by}{which column to rank the data on, supplied as variable.
default: zscore}

\item{descending}{logical, whether ranking should be done from highest to
lowest value. default: TRUE}

\item{recursive}{logical, whether directories should be scanned recursively
for pathway files. default: TRUE}
}
\value{
data frame with IPA pathways
}
\description{
Batch read IPA files into one data frame
}
