% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{cor_test_df}
\alias{cor_test_df}
\title{Wrapper for cor.test for data frames}
\usage{
cor_test_df(df, x, y)
}
\arguments{
\item{df}{data frame with numeric data to correlate, usually grouped by
another column}

\item{x}{numeric data frame colum, supplied as variable}

\item{y}{numeric data frame colum, supplied as variable}
}
\value{
tidy data frame of correlations
}
\description{
Wrapper for cor.test for data frames
}
