% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cp_pathways}
\alias{cp_pathways}
\title{A tibble with differentially regulated pathways identified by clusterProfiler}
\format{
tibble with 116 rows and 5 columns
\describe{
\item{ID}{Pathway ID, chr}
\item{Description}{Pathway name, chr}
\item{GeneRatio/NES}{GeneRatio for ORA or NES for GSEA, num}
\item{pvalue}{P-value for being differentially regulated, num}
\item{p.adjust}{Adjusted p-value for being differentially regulated, num}
}
}
\usage{
cp_pathways
}
\description{
A tibble with differentially regulated pathways identified by clusterProfiler
}
\keyword{datasets}
