% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{create_gsea_cls}
\alias{create_gsea_cls}
\title{Create class file for GSEA analysis}
\usage{
create_gsea_cls(
  data,
  metadata,
  out = NULL,
  sample_colname = sample,
  group_colname = group
)
}
\arguments{
\item{data}{data frame containing gene expression data with gene ID/symbol
column and sample names as columns, only samples found in metadata will be
used}

\item{metadata}{data frame containing sample and group information, requires
a column for samples and column for grouping}

\item{out}{character string, optional output filename. If NULL, only
formatted text is returned. default: NULL}

\item{sample_colname}{column name of the sample column, supplied as variable.
default: sample}

\item{group_colname}{column name of the sample column, supplied as variable.
default: group}
}
\value{
formatted text GSEA class file, if out is supplied, text will be
saved to the corresponding file
}
\description{
Create class file for GSEA analysis
}
