% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{create_gsea_normalized}
\alias{create_gsea_normalized}
\title{Create normalized gene expression file for GSEA analysis}
\usage{
create_gsea_normalized(
  data,
  metadata,
  id_colname,
  data_description_col = NULL,
  sample_colname = sample,
  out = NULL
)
}
\arguments{
\item{data}{data frame containing gene expression data with gene ID/symbol
column and sample names as columns, only samples found in metadata will be
used}

\item{metadata}{data frame containing sample and group information, requires
a column for samples and column for grouping}

\item{id_colname}{column name with gene IDs/symbols in the data file,
supplied as variable}

\item{data_description_col}{optional column name with gene descriptions
in the data file, supplied as variable. If NULL the resulting DESCRIPTION
column in the GSEA expression file will be empty. default: NULL}

\item{sample_colname}{columns column name of the sample column in the
metadata file, supplied as variable. default: sample}

\item{out}{character string, optional output filename. If NULL, only
formatted data frame is returned. default: NULL}
}
\value{
formatted data frame with normalized gene expression for  GSEA
analysis, if out is supplied, it will besaved to the corresponding file
}
\description{
Create normalized gene expression file for GSEA analysis
}
