% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{create_gsea_rank}
\alias{create_gsea_rank}
\title{Create gene ranking file for GSEA Preranked analysis}
\usage{
create_gsea_rank(
  data,
  out = NULL,
  ranking_formula = ~-log10(pvalue) * sign(log2FoldChange),
  id_colname = ensembl_gene_id,
  .inf = "replace"
)
}
\arguments{
\item{data}{data frame containing gene expression data with a gene ID/symbol
column}

\item{out}{character string, optional output filename. If NULL, only
formatted text is returned. default: NULL}

\item{ranking_formula}{formula, way to calculate ranking of genes,
downregulated genes should have negative ranking, formula shouldn't produce
duplicates}

\item{id_colname}{column name with gene IDs/symbols in the data file,
supplied as variable}

\item{.inf}{one of ("drop", "replace"). How Inf values should be treated in
the data. "replace" will replace the Inf values with values 1\% higher than
the current highest value}
}
\value{
formatted data frame with ranked genes for GSEA Preranked analysis,
if out is supplied, it will besaved to the corresponding file
}
\description{
Create gene ranking file for GSEA Preranked analysis
}
