% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diffexp_data}
\alias{diffexp_data}
\title{A list of tibbles with differentially expressed genes identified by DESeq2.
Each tibble corresponds to one comparison between groups.}
\format{
a list of tibbles of length 2 each with 19229 rows and 10 columns
\describe{
\item{ex_1}{group comparison 1}
\item{ENSEMBL}{ensembl gene id, chr}
\item{ENTREZID}{entrez id, num}
\item{SYMBOL}{gene symbol, chr}
\item{GENENAME}{long gene name, chr}
\item{baseMean}{mean of normalized counts for all samples, num}
\item{log2FoldChange}{log2 fold-change difference of expression, num}
\item{lfcSE}{standard error, num}
\item{stat}{Wald statistic, num}
\item{pvalue}{P-value for differential expression, num}
\item{padj}{Adjusted p-value for differential expression, num}
\item{ex_2}{group comparison 2}
\item{ENSEMBL}{ensembl gene id, chr}
\item{ENTREZID}{entrez id, num}
\item{SYMBOL}{gene symbol, chr}
\item{GENENAME}{long gene name, chr}
\item{baseMean}{mean of normalized counts for all samples, num}
\item{log2FoldChange}{log2 fold-change difference of expression, num}
\item{lfcSE}{standard error, num}
\item{stat}{Wald statistic, num}
\item{pvalue}{P-value for differential expression, num}
\item{padj}{Adjusted p-value for differential expression, num}
}
}
\usage{
diffexp_data
}
\description{
A list of tibbles with differentially expressed genes identified by DESeq2.
Each tibble corresponds to one comparison between groups.
}
\keyword{datasets}
