% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dir_exists}
\alias{dir_exists}
\title{Check if directory exists}
\usage{
dir_exists(dir_string)
}
\arguments{
\item{dir_string}{character string of path to check}
}
\value{
character string of path if it exists, stops executions otherwise
}
\description{
Check if directory exists
}
