% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dire}
\alias{dire}
\title{A tibble with upstream transcription factors (TFs) identified by dire.dcode.org}
\format{
tibble with 241 rows and 4 columns
\describe{
\item{Transcription Factor}{transcription factor name, chr}
\item{Occurrence}{TF occurrence, num}
\item{Importance}{TF importance, num}
\item{file}{name of the file the data was parsed from, chr}
}
}
\usage{
dire
}
\description{
A tibble with upstream transcription factors (TFs) identified by dire.dcode.org
}
\keyword{datasets}
