% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{.addFeatures}
\alias{.addFeatures}
\title{Upgrade of addFeatures from cummeRbund package}
\usage{
.addFeatures(object, features, level = "genes", ...)
}
\arguments{
\item{object}{aaa}

\item{features}{bbb}

\item{level}{ccc}

\item{...}{ddd}
}
\value{
dbi
}
\description{
It was using some deprected function
}
