% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expression_data}
\alias{expression_data}
\title{A tibble with normalized gene counts for all samples obtained from DESeq2}
\format{
tibble with 19229 rows and 19 columns
\describe{
\item{ENSEMBL}{ensembl gene id, chr}
\item{ENTREZID}{entrez id, num}
\item{SYMBOL}{gene symbol, chr}
\item{GENENAME}{long gene name, chr}
\item{cgrp1}{normalized counts for sample cgrp1, num}
\item{cgrp2}{normalized counts for sample cgrp2, num}
\item{cgrp3}{normalized counts for sample cgrp3, num}
\item{ctrl1}{normalized counts for sample ctrl1, num}
\item{ctrl2}{normalized counts for sample ctrl2, num}
\item{ctrl3}{normalized counts for sample ctrl3, num}
\item{ne1}{normalized counts for sample ne1, num}
\item{ne2}{normalized counts for sample ne2, num}
\item{ne3}{normalized counts for sample ne3, num}
\item{npy1}{normalized counts for sample npy1, num}
\item{npy2}{normalized counts for sample npy2, num}
\item{npy3}{normalized counts for sample npy3, num}
\item{sp1}{normalized counts for sample sp1, num}
\item{sp2}{normalized counts for sample sp2, num}
\item{sp3}{normalized counts for sample sp3, num}
}
}
\usage{
expression_data
}
\description{
A tibble with normalized gene counts for all samples obtained from DESeq2
}
\keyword{datasets}
