% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{filter_string}
\alias{filter_string}
\title{Filter column values based on regex}
\usage{
filter_string(df, colname, str)
}
\arguments{
\item{df}{data frame to filter}

\item{colname}{character string column name of values to filter}

\item{str}{regex to filter by}
}
\value{
data frame
}
\description{
Filter column values based on regex
}
