% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_lists}
\alias{gene_lists}
\title{A named list of character vectors containing user-defined gene lists parsed
from json file}
\format{
named list with 6 items
\describe{
\item{mito_1}{mitchondrial genes, chr vector}
\item{metabolism}{metabolism genes, chr vector}
\item{volcano_example}{gene lists for example volcano plot, chr vector}
\item{heatmap_fc_example_1}{gene lists for example heatmap, chr vector}
\item{heatmap_fc_example_2}{gene lists for example heatmap, chr vector}
\item{oxhphos}{oxphos genes, chr vector}
}
}
\usage{
gene_lists
}
\description{
A named list of character vectors containing user-defined gene lists parsed
from json file
}
\keyword{datasets}
