% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{get_biomart_gene_mapping}
\alias{get_biomart_gene_mapping}
\title{Get gene ID mapping from biomart}
\usage{
get_biomart_gene_mapping(mart, id_attribute, mapped_attribute, ...)
}
\arguments{
\item{mart}{biomart mart to use}

\item{id_attribute}{character string, source attribute name}

\item{mapped_attribute}{character vector, names of target attributes}

\item{...}{other parameters to biomaRt getBM function}
}
\value{
data frame of id_attributed IDs mapped to mapped_attribute IDs
}
\description{
An alternative to clusterProfiler::bitr function
}
