% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{get_plot_params}
\alias{get_plot_params}
\title{Return export parameters for a plot}
\usage{
get_plot_params(type, plot_params)
}
\arguments{
\item{type}{character string, type of plots, corresponds to the graph_type
property in the config file}

\item{plot_params}{data frame or a character string pointing to a json file
with plot export parameters}
}
\value{
named list of plot export parameters
}
\description{
Paramters will contain properties such as filename, export device, dpi,
units, width and height of plot
}
