% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{get_sig_genes}
\alias{get_sig_genes}
\title{Get significantly regulated genes in character vector}
\usage{
get_sig_genes(
  diffexp_data,
  id_colname = SYMBOL,
  filter_sig_on = padj,
  sig_threshold = 0.05
)
}
\arguments{
\item{diffexp_data}{data frame of gene differential expression, must contain
column for gene IDs/symbols and a p-value}

\item{id_colname}{column name of gene IDs/symbols in differential expression
data file, supplied as variable}

\item{filter_sig_on}{usually column name of p-values in differential
expression data file values of which will determine significance,
supplied as variable. default: padj}

\item{sig_threshold}{double, significance threshold for filter_sig_on column}
}
\value{
character vector of significant gene IDs/symbols
}
\description{
Get significantly regulated genes in character vector
}
