% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Create default ggplot2 color hues}
\usage{
gg_color_hue(i)
}
\arguments{
\item{i}{integer, number of colors to generate}
}
\value{
character vector with equally spaced color hues
}
\description{
Create default ggplot2 color hues
}
