% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{ggsave_fixed}
\alias{ggsave_fixed}
\title{Save ggplot graph with fixed panel sizes}
\usage{
ggsave_fixed(
  file,
  plot = last_plot(),
  units = "mm",
  margin = 1,
  panel_width = 4,
  panel_height = 4,
  width = round(dev.size()[1], digits = 1),
  height = round(dev.size()[1], digits = 1)
)
}
\arguments{
\item{file}{character string, file name of the exported file}

\item{plot}{plot to save, defaults to last plot}

\item{units}{character string, plot size units, one of
("in", "cm", "mm", or "px"), default: "mm"}

\item{margin}{double, plot panel size in specified units, default: 1}

\item{panel_width}{double, plot panel width in specified units, default: 4}

\item{panel_height}{double, plot panel height in specified units, default: 4}

\item{width}{double, plot width in specified units}

\item{height}{double, plot height in specified units}
}
\value{
ggplot graph with fixed panel sizes
}
\description{
Useful when the axis labels between group of graphs are of different sizes
and you want to keep the plotting areas of equal dimensions
}
