% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{ggsave_param}
\alias{ggsave_param}
\title{Save ggplot graph with specified parameters}
\usage{
ggsave_param(
  output_dir,
  plot_params,
  plot = last_plot(),
  filename_prefix = "",
  filename_suffix = "",
  date_prefix = FALSE
)
}
\arguments{
\item{output_dir}{character string, path to ouput directory, filename is
specified in plot_params}

\item{plot_params}{named list of export parameters for ggsave function such
as width, height, dpi, device etc. Includes the file name of the plot}

\item{plot}{plot to export. default: last plot}

\item{filename_prefix}{character string, add this as a prefix to the file
name}

\item{filename_suffix}{character string, add this as a suffix to the file
name}

\item{date_prefix}{logical, whether to include current date at the beginning
of the plot name}
}
\description{
Parameters would be ideally supplied by a config json file
}
