% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_config}
\alias{load_config}
\title{Parse config file}
\usage{
load_config(x, into = "list", ...)
}
\arguments{
\item{x}{charcter string, list or environment to parse config from.
Character string has to be a location to a json file}

\item{into}{character string, one of c("list", "env") whether to return
loaded config file as list or as an environment}

\item{...}{other parameters passed to load_config}
}
\value{
list or environment of parsed config file
}
\description{
Parse config file
}
