% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data_diffexp}
\alias{load_data_diffexp}
\title{Load data containing differetial expression}
\usage{
load_data_diffexp(path, params)
}
\arguments{
\item{path}{named list, names have to be unique, values are character strings
corresponding to file paths of data with differential expression}

\item{params}{list of parameters parsed from config file. Currently unused}
}
\value{
data frame
}
\description{
Wrapper function for loading data that exposes common interface
consisting of path to file/directory and parameter list parsed from json
config file
}
