% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_gene_lists}
\alias{load_gene_lists}
\title{Load user specified gene lists from json file}
\usage{
load_gene_lists(path, params)
}
\arguments{
\item{path}{character string, path to json gene list file}

\item{params}{list of parameters parsed from config file. Currently unused}
}
\value{
named list, where names correspond to gene list names supplied by
user and values are character vectors of gene symbols/IDs
}
\description{
Wrapper function for loading data that exposes common interface
consisting of path to file/directory and parameter list parsed from json
config file. Gene list json file has user specified names and values can be
either lists of lenght > 1 containing gene symbols or lists of length == 1
containing path to text file with gene list to be read. File based gene lists
contain single column of gene names without header.
}
