% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_into}
\alias{load_into}
\title{Helper for load config into list or environment}
\usage{
load_into(x, into)
}
\arguments{
\item{x}{named list, config file parsed from json or other source}

\item{into}{character string, one of c("list", "env") whether to return
loaded config file as list or as an environment}
}
\value{
list or environment of parsed config file
}
\description{
Helper for load config into list or environment
}
