% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_metadata}
\alias{load_metadata}
\title{Load metadata information}
\usage{
load_metadata(path, params)
}
\arguments{
\item{path}{character string, path to metadata file}

\item{params}{list of parameters parsed from config file, requires
sample_colname, group_colname and group_levels to be present}
}
\value{
data frame
}
\description{
Wrapper function for loading data that exposes common interface
consisting of path to file/directory and parameter list parsed from json
config file
}
