% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_pathway_lists}
\alias{load_pathway_lists}
\title{Load user specified pathway lists from json file}
\usage{
load_pathway_lists(path, params)
}
\arguments{
\item{path}{character string, path to json pathway list file}

\item{params}{list of parameters parsed from config file. Currently unused}
}
\value{
named list, where names correspond to gene list names supplied by
user and values are integer or character vectors
}
\description{
Wrapper function for loading data that exposes common interface consisting
of path to file/directory and parameter list parsed from json config file.
Gene list json file has user specified names and values can be either
integer lists corresponding to the rank of the pathway in data or character
list with names of pathways.
}
