% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_pathways_amigo}
\alias{load_pathways_amigo}
\title{Load data from AmiGO analysis}
\usage{
load_pathways_amigo(path, params)
}
\arguments{
\item{path}{character string, path to directory with AmiGO analysis.
Will search the path recursively}

\item{params}{list of parameters parsed from config file. Requires
amigo_pathways_pattern and amigo_filter_level_up_to to be present}
}
\value{
data frame
}
\description{
Wrapper function for loading data that exposes common interface
consisting of path to file/directory and parameter list parsed from json
config file. Accepts data from AmiGO analysis in the form of json files
due to the preservation of hierarchical pathway analysis.
}
