% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_pathways_dire}
\alias{load_pathways_dire}
\title{Load dire analysis files}
\usage{
load_pathways_dire(path, params)
}
\arguments{
\item{path}{character string, path to directory containing files with dire
analysis. Will search the path recursively, see collate_pathways_dire for
more information}

\item{params}{list of parameters parsed from config file. Requires
dire_pathways_pattern and dire_sheet_name to be present}
}
\value{
data frame
}
\description{
Wrapper function for loading data that exposes common interface
consisting of path to file/directory and parameter list parsed from json
config file
}
