% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_pathways_gsea}
\alias{load_pathways_gsea}
\title{Load data from GSEA analysis}
\usage{
load_pathways_gsea(path, params)
}
\arguments{
\item{path}{character string, path to directory with GSEA analysis.
Will search the path recursively, see collate_pathways_gsea for more
information}

\item{params}{list of parameters parsed from config file. Requires
gsea_pathways_pattern and pvalue_threshold to be present}
}
\value{
data frame
}
\description{
Wrapper function for loading data that exposes common interface
consisting of path to file/directory and parameter list parsed from json
config file. Accepts data from GUI GSEA app.
}
