% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_pathways_ipa}
\alias{load_pathways_ipa}
\title{Load data from IPA analysis}
\usage{
load_pathways_ipa(path, params)
}
\arguments{
\item{path}{character string, path to directory with IPA analysis.
Will search the path recursively}

\item{params}{list of parameters parsed from config file. Requires
ipa_pathways_pattern and ipa_rank_by to be present. ipa_rank_by denotes the
column name to rank the pathways by}
}
\value{
data frame
}
\description{
Wrapper function for loading data that exposes common interface
consisting of path to file/directory and parameter list parsed from json
config file. Accepts data from Qiagen IPA analysis that were processed by
github.com/icervenka/ipa_reports_snakemake.
}
