% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{A tibble with sample-group mapping}
\format{
tibble with 15 rows and 2 columns
\describe{
\item{sample}{sample name/id, chr}
\item{group}{group name, chr}
}
}
\usage{
metadata
}
\description{
A tibble with sample-group mapping
}
\keyword{datasets}
