% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mito_atp_replacer}
\alias{mito_atp_replacer}
\title{A named list used for replacing gene names. Names contain gene names to be
replaced and values contain new gene names}
\format{
a named list of length 13
}
\usage{
mito_atp_replacer
}
\description{
A named list used for replacing gene names. Names contain gene names to be
replaced and values contain new gene names
}
\keyword{datasets}
