% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modify_gene_symbols}
\alias{modify_gene_symbols}
\title{Text processing of gene symbols}
\usage{
modify_gene_symbols(gene_list, text_replacements, text_modifications)
}
\arguments{
\item{gene_list}{character vector of gene symbols to process}

\item{text_replacements}{named vector, where names are patterns and values
replacement. Will be applied in order it is specified.}

\item{text_modifications}{list of single argument text processing functions
to apply to gene_list. Will be applied from right to left.}
}
\value{
character vector of modified gene names
}
\description{
Text processing of gene symbols
}
