% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mouseify_gene_symbols}
\alias{mouseify_gene_symbols}
\title{Replace mouse gene symbols}
\usage{
mouseify_gene_symbols(gene_list)
}
\arguments{
\item{gene_list}{character vector of gene symbols to process}
}
\value{
character vector of modified gene names
}
\description{
Applies simple text replacements for mitochondrial gene symbols  and converts
gene symbols to title case. This is a more specific version of
modify_gene_symbols function and a fast way to replace human gene symbols
with reasonable accuracy.
}
