% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathway_lists}
\alias{pathway_lists}
\title{A named list of either numeric or character vectors containing pathway IDs or
names to filter from a tibble containing differentially regulated pathways}
\format{
a named list of length 2
\describe{
\item{example_1}{example pathways to filter, num vector}
\item{example_volcano}{example pathways to filter for volcano plot, num vector}
}
}
\usage{
pathway_lists
}
\description{
A named list of either numeric or character vectors containing pathway IDs or
names to filter from a tibble containing differentially regulated pathways
}
\keyword{datasets}
