% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.R
\name{plot_cp_pathways_bargraph}
\alias{plot_cp_pathways_bargraph}
\title{Simple barplot of differentally regulated pathways}
\usage{
plot_cp_pathways_bargraph(
  pathway_data,
  pathway_source_pattern,
  top_pathways = 20,
  truncate_name = 80
)
}
\arguments{
\item{pathway_data}{data frame with data for pathway analysis. Should include
a columns for enrichment, p-value, pathway name and sorted based on their
rank.}

\item{pathway_source_pattern}{character string, filter for a specific pathway
source such as Gene Ontology, WikiPathways. See
clusterprofiler_reports_snakemake output file to see the possible values.}

\item{top_pathways}{integer, top n pathways to display. default: 20}

\item{truncate_name}{integer, maximum number of characters after which the
pathway names will be truncated. default: 80}
}
\value{
ggplot barplot
}
\description{
Uses data from clusterprofiler_reports_snakemake analysis.
}
