% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.R
\name{plot_cp_pathways_meta}
\alias{plot_cp_pathways_meta}
\title{Meta analysis of pathway data}
\usage{
plot_cp_pathways_meta(pathway_data, top_pathways = 30)
}
\arguments{
\item{pathway_data}{data frame with data for pathway analysis. Should include
a columns for enrichment, p-value, pathway name and sorted based on their
rank. Usually it would be the whole collated output from
clusterprofiler_pipline.}

\item{top_pathways}{integer, number of top and bottom patways to plot.
default: 30}
}
\value{
grid of 3 ggplot plots
}
\description{
This function is used to plot the overview of differentially regulated
pathway sets from the  clusterprofiler_pipline workflow. Creates a grid of
3 plots: i) histogram of number of differentially regulated pathways from
pathways databases, ii) databases with highest amount of differentially
regulated pathways, iii) databases with lowest amount of differentially
regulated pathways.
}
