% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dire.R
\name{plot_dire}
\alias{plot_dire}
\title{Creates plot from dire.dcode.com upstream transcription factor data}
\usage{
plot_dire(df, dot_size = 2.5, color = "steelblue4")
}
\arguments{
\item{df}{data frame containing column of transcription factor names,
Occurrence and Importance columns}

\item{dot_size}{size of scatter plot symbols}

\item{color}{color of scatter plot symbols}
}
\value{
ggplot scatter plot
}
\description{
Creates plot from dire.dcode.com upstream transcription factor data
}
