% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dire.R
\name{plot_dire_labeled}
\alias{plot_dire_labeled}
\title{Creates labeled plot from dire.dcode.com upstream transcription factor data}
\usage{
plot_dire_labeled(
  df,
  occurrence_threshold = 0.05,
  importance_threshold = 0.05,
  gene_list = NULL,
  combine_thresholds = `&`,
  ...
)
}
\arguments{
\item{df}{data frame containing column of transcription factor names,
Occurence and Importance columns}

\item{occurrence_threshold}{double, minimum occurrence value to plot gene
labels}

\item{importance_threshold}{double, minimum importance value to plot gene
labels}

\item{gene_list}{character vector, label specific set of genes, overrides
occurrence and importance thresholds}

\item{combine_thresholds}{binary logical operator, how to combine the
occurrence and importance thresholds, default '&'}

\item{...}{other parameters to plot_dire function}
}
\value{
ggplot scatter plot with labeled transcription factors of interest
}
\description{
Creates labeled plot from dire.dcode.com upstream transcription factor data
}
