% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Draws a heatmap of gene expression}
\usage{
plot_heatmap(
  expression_data,
  metadata,
  gene_list = NULL,
  include_groups = NULL,
  id_colname = SYMBOL,
  process_gene_dup_fun = sum,
  sample_colname = sample,
  group_colname = group,
  gene_ranking_fun = matrixStats::rowVars,
  cell_dims = c(10, 1),
  palette = "RdBu",
  ...
)
}
\arguments{
\item{expression_data}{data frame of normalized gene expression with samples
as columns and an id column for gene names/symbols}

\item{metadata}{data frame with column of samples names, column for group
names and optionally other columns describing the sample features}

\item{gene_list}{can be one of:
\itemize{
\item character vector of gene IDs/symbols to display
\item integer, top ranking genes based on gene_ranking_fun will be displayed
\item NULL, all genes will be displayed
}}

\item{include_groups}{character vector, which groups from metadata file to
include. If NULL, all groups will be included. default: NULL}

\item{id_colname}{column name of gene IDs/symbols in expression data file,
supplied as variable. default: SYMBOL}

\item{process_gene_dup_fun}{function, how to treat the expression data for
potential gene ID/symbol duplicates. default: sum}

\item{sample_colname}{column name of samples in metadata file, supplied as
variable. default: sample}

\item{group_colname}{column name of experimental groups in metadata file,
supplied as variable. default: group}

\item{gene_ranking_fun}{function, in case of displaying top genes, how should
the genes in the matrix be ranked. default: matrixStats::rowVars}

\item{cell_dims}{interger vector of length 2 specifying width and height of
individual heatmap cells. default: c(10, 1)}

\item{palette}{character, which colorbrewer palette to use. default: "RdBu"}

\item{...}{other parameters to pheatmap function}
}
\value{
pheatmap plot
}
\description{
Uses pheatmap package.
}
