% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{plot_heatmap_fc}
\alias{plot_heatmap_fc}
\title{Creates a heatmap combining expression and differential expression data}
\usage{
plot_heatmap_fc(
  expression_data,
  diffexp_data,
  metadata,
  gene_list,
  include_groups = NULL,
  id_colname = SYMBOL,
  fc_colname = log2FoldChange,
  .fc_colors = c("darkred", "steelblue"),
  pval_colname = padj,
  pval_threshold = 0.05,
  .pval_colors = c("white", "steelblue"),
  sample_colname = sample,
  group_colname = group,
  .col_annot_colors = NULL,
  ...
)
}
\arguments{
\item{expression_data}{data frame of normalized gene expression with samples
as columns and an id column for gene names/symbols}

\item{diffexp_data}{data frame of gene differential expression, must contain
column for gene IDs/symbols, log2 fold-changes and a p-value}

\item{metadata}{data frame with column of samples names, column for group
names and optionally other columns describing the sample features}

\item{gene_list}{character vector of gene IDs/symbols to display}

\item{include_groups}{character vector, which groups from metadata file to
include. If NULL, all groups will be included. default: NULL}

\item{id_colname}{column name of gene IDs/symbols in expression data file,
supplied as variable. default: SYMBOL}

\item{fc_colname}{column name of log fold-changes in the data frame with
differential expression data, supplied as variable. default: log2FolcChange}

\item{.fc_colors}{character vector of length 2, colors that will be mapped to
highest and lowest log fold-changes respectively. default:
c("darkred", "steelblue")}

\item{pval_colname}{column name of p-values or adjusted p-values in the
data frame with differential expression data, supplied as variable.
default: padj}

\item{pval_threshold}{double, significance threshold for p-values, used for
mapping the lower value from .pval_colors to this value. default: 0.05}

\item{.pval_colors}{character vector of length 2, colors that will be mapped
to highest and lowest p-values respectively. default: c("white", "steelblue")}

\item{sample_colname}{column name of samples in metadata file, supplied as
variable. default: sample}

\item{group_colname}{column name of experimental groups in metadata file,
supplied as variable. default: group}

\item{.col_annot_colors}{names character vector of length corresponding to
the number of displayed groups. Names of vector correspond group names and
values to their annotation colors. If NULL, standard ggplot color scheme
will be used. default: NULL}

\item{...}{extra params to ComplexHeatmap}
}
\value{
ComplexHeatmap heatmap
}
\description{
Heatmap will have a plot for p-values and log2 fold-changes of selected
comparison on the right side. Uses ComplexHeatmap.
}
