% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma.R
\name{plot_ma}
\alias{plot_ma}
\title{Draws MA-plot based on differential expression data}
\usage{
plot_ma(
  diffexp_data,
  x = baseMean,
  y = log2FoldChange,
  id_colname = SYMBOL,
  filter_sig_on = padj,
  sig_threshold = 0.05,
  x_threshold = 0,
  y_threshold = 0,
  label_bottom_n = 5,
  label_top_n = 5,
  gene_list = NULL,
  xlab_label = expression(paste("lo", g[2], "Expression")),
  ylab_label = expression(paste("lo", g[2], "FC")),
  color_palette = c("gray70", "steelblue", "darkred")
)
}
\arguments{
\item{diffexp_data}{data frame of gene differential expression, must contain
column for gene IDs/symbols, log2 fold-changes and a p-value}

\item{x}{column name in diffexp_data to be plotted on the x axis, supplied
as variable. For MA plots, log2(x) is plotted. default: baseMean}

\item{y}{column name in diffexp_data to be plotted on the y axis, supplied
as variable. default: log2FoldChange}

\item{id_colname}{column name of gene IDs/symbols in differential expression
data file, supplied as variable}

\item{filter_sig_on}{usually column name of p-values in differential
expression data file values of which will determine significance,
supplied as variable. default: padj}

\item{sig_threshold}{double, significance threshold for filter_sig_on column}

\item{x_threshold}{double, significance threshold for x column. default: 0}

\item{y_threshold}{double, significance threshold for y column. default: 0}

\item{label_bottom_n}{integer, number of labels to display for genes with
highest y value. default: 5}

\item{label_top_n}{integer, number of labels to display for genes with
lowest y value. default: 5}

\item{gene_list}{character vector of gene IDs/symbols to display, overrides
label_top_n and label_bottom_n settings. default: NULL}

\item{xlab_label}{character string, label for x axis. default: log2Expression}

\item{ylab_label}{character string, label for x axis. default: log2FC}

\item{color_palette}{character vector of color palette, length has to be
equal to 3 for unchaged, downregulated and upregulated genes. default:
c("gray70", "steelblue","darkred")}
}
\value{
MA plot with colored significat genes and optionally gene labels
}
\description{
Adds labels for genes of interest and colors based on upregulated and
downregulated genes.
}
