% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{plot_param_corr}
\alias{plot_param_corr}
\title{Create a correlation graph of gene expression and sample parameter}
\usage{
plot_param_corr(
  expression_data,
  metadata,
  gene_list,
  params,
  id_colname = SYMBOL,
  sample_colname = sample,
  group_colname = group,
  palette = NULL,
  show_regression = TRUE
)
}
\arguments{
\item{expression_data}{data frame of normalized gene expression with samples
as columns and an id column for gene names/symbols}

\item{metadata}{data frame with column of samples names, column for group
names and optionally other columns describing the sample features that will
be correlated with gene expression}

\item{gene_list}{character vector of gene IDs/symbols, which will be used
to  plot the correlations}

\item{params}{character vector stating which parameter columns to choose
from metadata to plot the  gene expression}

\item{id_colname}{column name of gene IDs/symbols in expression data file,
supplied as variable}

\item{sample_colname}{column name of samples in metadata file, supplied as
variable. default: sample}

\item{group_colname}{column name of experimental groups in metadata file,
supplied as variable. default: group}

\item{palette}{character vector of color palette, length has to be equal to
the amount of groups in metadata file}

\item{show_regression}{logical, whether to show regression line on the graph}
}
\value{
ggplot facet graph of correlation between specified sample parameters
and gene expression of selected genes
}
\description{
Create a correlation graph of gene expression and sample parameter
}
