% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.R
\name{plot_pathways_volcano}
\alias{plot_pathways_volcano}
\title{Create volcano plot for identified pathways}
\usage{
plot_pathways_volcano(
  pathway_data,
  x,
  y,
  label = name,
  x_threshold = 0,
  y_threshold = 0,
  label_pathways = NULL,
  xlab_text = "Normalized Enrichment Score",
  ylab_text = "-log10(FDR)",
  add_vhlines = TRUE,
  vhline_color = "gray40",
  vhline_type = "dashed",
  alpha = 0.2,
  color_palette = c("gray70", "steelblue4", "darkred")
)
}
\arguments{
\item{pathway_data}{data frame with data for pathway analysis. Should include
a columns for enrichment, p-value and a pathway name or ID.}

\item{x}{What to plot on x axis, supplied as variable present as
a column in pathway_data variable.}

\item{y}{What to plot on y axis, supplied as variable present as
a column in pathway_data variable.}

\item{label}{Which column in the pathway_data contains names, supplied as
variable.}

\item{x_threshold}{double, significance threshold for x axis.}

\item{y_threshold}{double, significance threshold for y axis.}

\item{label_pathways}{charater or integer vector. Which pathways to include
in the graph. If character vector, pathways are selected by names, if integer
vector, pathways are selected by their rank.}

\item{xlab_text}{character string, text label for x axis. default:
"Normalized Enrichment Score"}

\item{ylab_text}{character string, text label for y axis. default:
"-log10(FDR)"}

\item{add_vhlines}{logical, whether to include horizontal a vertical lines
that show the significance threholds. default: TRUE}

\item{vhline_color}{character string, color of the vertical and horizontal
lines. default: grey40}

\item{vhline_type}{character string, line type of vertical and horizonal.
default: dashed}

\item{alpha}{double, transparency of the geom_point.}

\item{color_palette}{character vector of length 3, colors for the unchaged,
down- and upregulated pathways.}
}
\value{
ggplot pathway volcano plot
}
\description{
Works with pathway analysis that supplies a measure of overrepresentation
with directionality, such as GSEA.
}
