% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_pca_common}
\alias{plot_pca_common}
\title{Draw PCA plot from normalized gene counts}
\usage{
plot_pca_common(expression_data, metadata, ...)
}
\arguments{
\item{expression_data}{data frame of normalized gene expression with samples
as columns and an id column for gene names/symbols}

\item{metadata}{data frame with column of samples names, column for group
names and optionally other columns describing the sample features}

\item{...}{other parameters to the plot_pca function, see ?plot_pca for more
information}
}
\value{
ggplot PCA plot
}
\description{
Draw PCA plot from normalized gene counts
}
