% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_pca_deseq}
\alias{plot_pca_deseq}
\title{Draw PCA plot from DESeq2 dds data}
\usage{
plot_pca_deseq(dds, norm = "vst", ...)
}
\arguments{
\item{dds}{DESeq2 dds file}

\item{norm}{character string, one of "vst" or "rlog". Type of count
normalization to be performed. default: vst.}

\item{...}{other parameters to the plot_pca function, see ?plot_pca for more
information}
}
\value{
ggplot PCA plot
}
\description{
Adds optional group centers and palette.
}
