% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{plot_venn}
\alias{plot_venn}
\title{Wrapper function for venn.diagram}
\usage{
plot_venn(
  l,
  output_directory,
  plot_params,
  category_names = "",
  palette = viridis::viridis(2),
  font_size = 0.6,
  ...
)
}
\arguments{
\item{l}{list of character vectors to plot in venn diagram.}

\item{output_directory}{character string. Path to the output directory where
the plots will be saved}

\item{plot_params}{list of output parameters for plot such as size, dpi,
units, file format, file name for export.}

\item{category_names}{character vector the same length as l containing
category names.}

\item{palette}{character vector the same length as l containg colors to use
as fills for categories.}

\item{font_size}{integer, font size for labels inside the venn diagram.}

\item{...}{other parameters to VennDiagram::venn.diagram function.}
}
\description{
Sets some reasonable defaults for looks. If plot_params$filename is specified
venn diagram is saved to this location, otherwise it is just drawn on
screen.
}
