% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rank_how}
\alias{rank_how}
\title{Helper switch logical toggle for descending ranking function}
\usage{
rank_how(descending)
}
\arguments{
\item{descending}{logical, whether to rank data frame columns in descending
order}
}
\value{
function  for ranking data frame entries
}
\description{
Helper switch logical toggle for descending ranking function
}
