% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_amigo_json}
\alias{read_amigo_json}
\title{Parse json file generated by AmiGO analysis}
\usage{
read_amigo_json(jsonfile)
}
\arguments{
\item{jsonfile}{json like list, generated by rjson::fromJSON function.}
}
\value{
data frame
}
\description{
Jsonfile generated by http://amigo.geneontology.org/amigo preserves pathway
hierarchy. Resulting data frame will have a level column indicating
hierarchy level.
}
