% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_data_cuffdiff}
\alias{read_data_cuffdiff}
\title{Read cufdiff data}
\usage{
read_data_cuffdiff(dir)
}
\arguments{
\item{dir}{character string, directory path that contains cuffdiff data.}
}
\value{
data frame
}
\description{
Read cufdiff data
}
