% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_dire_xlsx}
\alias{read_dire_xlsx}
\title{Read dire data from excel file}
\usage{
read_dire_xlsx(filename, sheet_name = "Sheet1")
}
\arguments{
\item{filename}{character string, file to read.}

\item{sheet_name}{character string, sheet name that contains the data.
default: Sheet1}
}
\value{
data frame
}
\description{
Result data from dire.dcode can be copied to excel while preserving tabular
formatting. The data should be kept as is from the webpage, the function
will take care of proper formatting.
}
