% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_gene_list_from_file}
\alias{read_gene_list_from_file}
\title{Read gene list from single column text file}
\usage{
read_gene_list_from_file(filename)
}
\arguments{
\item{filename}{text file to read that contains single column with gene names
or IDs.}
}
\value{
character vector
}
\description{
File shouldn't have a header.
}
