% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_pathways_amigo}
\alias{read_pathways_amigo}
\title{Parse json file generated by AmiGO analysis}
\usage{
read_pathways_amigo(filename)
}
\arguments{
\item{filename}{character vector, path to json file containing pathway
analysis from AmiGO}
}
\value{
data frame
}
\description{
Jsonfile generated by http://amigo.geneontology.org/amigo preserves pathway
hierarchy. Resulting data frame will have a level column indicating
hierarchy level.
}
