% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_pathways_cp}
\alias{read_pathways_cp}
\title{Read pathways generated by clusterprofiler_reports_snakemake}
\usage{
read_pathways_cp(filename, rank_by = `GeneRatio/NES`, descending = TRUE)
}
\arguments{
\item{filename}{character string, text file with clusterProfiler
pathways to process}

\item{rank_by}{which column should the data be ranked on, supplied as
variable. default: \code{GeneRatio/NES}}

\item{descending}{logical, whethere the renking should be from highest to
lowest}
}
\value{
data frame
}
\description{
See https://github.com/icervenka/clusterprofiler_reports_snakemake for more
information
}
