% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_pathways_gsea}
\alias{read_pathways_gsea}
\title{Read GSEA result file}
\usage{
read_pathways_gsea(
  filename,
  rank_by = NES,
  descending = TRUE,
  rename_columns = c("name", "link", "details", "size", "es", "nes", "nom_pval",
    "fdr_qval", "fwer_pval", "rank_at_max", "leading_edge")
)
}
\arguments{
\item{filename}{character string, text file with GSEA pathways to process}

\item{rank_by}{which column should the data be ranked on, supplied as
variable. default: NES}

\item{descending}{logical, whethere the renking should be from highest to
lowest}

\item{rename_columns}{character vector of size 11, new names for data frame
columns. Applied as a last operation. default: c("name", "link", "details",
"size", "es", "nes", "nom_pval", "fdr_qval", "fwer_pval", "rank_at_max",
"leading_edge")}
}
\value{
data frame
}
\description{
Adds a filename column containing the filename parameter.
}
