% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqplots-package.R
\docType{package}
\name{rnaseqplots-package}
\alias{rnaseqplots}
\alias{rnaseqplots-package}
\title{rnaseqplots: Create publication-ready plots from RNA-Seq data}
\description{
Use data for differentially expressed genes and pathways as well as normalized gene counts to easily create standard plots for RNA-Seq data such as MA, PCA, Volcano, Heatmaps etc. and export them for publications.
}
\author{
\strong{Maintainer}: Igor Cervenka \email{igor.cervenka@gmail.com} (\href{https://orcid.org/0000-0002-9438-5161}{ORCID})

}
\keyword{internal}
