% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_config}
\alias{validate_config}
\title{Checks is item is present in a configuration dictionary}
\usage{
validate_config(entry, config_dict)
}
\arguments{
\item{entry}{character string to check}

\item{config_dict}{named list usually obtained by parsing json configuration
file}
}
\value{
TRUE or FALSE depending whether the entry is present as a name in the
list
}
\description{
Checks is item is present in a configuration dictionary
}
