% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Create an annotated volcano plot}
\usage{
volcano_plot(
  diffexp_data,
  x = log2FoldChange,
  y = pvalue,
  id_colname = SYMBOL,
  filter_sig_on = padj,
  sig_threshold = 0.05,
  log2fc_threshold = 0.585,
  label_bottom_n = 5,
  label_top_n = 5,
  gene_list = NULL,
  add_vhlines = TRUE,
  vhline_color = "gray40",
  vhline_type = "dashed",
  xlab_label = expression(paste("lo", g[2], "FC")),
  ylab_label = expression(paste("-lo", g[10], "FDR")),
  color_palette = c("gray70", "steelblue", "darkred")
)
}
\arguments{
\item{diffexp_data}{data frame of gene differential expression, must contain
column for gene IDs/symbols, log2 fold-changes and a p-value}

\item{x}{column name in diffexp_data to be plotted on the x axis, supplied
as variable. default: log2FoldChange}

\item{y}{column name in diffexp_data to be plotted on the y axis, supplied
as variable. default: pvalue}

\item{id_colname}{column name of gene IDs/symbols in differential expression
data file, supplied as variable. default: SYMBOL}

\item{filter_sig_on}{usually column name of p-values in differential
expression data file values of which will determine significance,
supplied as variable. default: padj}

\item{sig_threshold}{double, significance threshold for filter_sig_on column}

\item{label_bottom_n}{integer, number of labels to display for genes with
highest y value. default: 5}

\item{label_top_n}{integer, number of labels to display for genes with
lowest y value. default: 5}

\item{gene_list}{character vector of gene IDs/symbols to display, overrides
label_top_n and label_bottom_n settings. default: NULL}

\item{add_vhlines}{logical, wheteher to add vertical and horizontal lines
indicating significance levels for both axes. default: TRUE}

\item{vhline_color}{character string, color of vertical and horizontal lines.
default: grey40}

\item{vhline_type}{character string, line type of vertical and horizontal
lines. Lines types for ggplot geom_line are accepted. default: dashed}

\item{xlab_label}{character string, label for x axis. default: log2FC}

\item{ylab_label}{character string, label for x axis. default: -log10FDR}

\item{color_palette}{character vector of color palette, length has to be
equal to 3 for unchaged, downregulated and upregulated genes. default:
c("gray70", "steelblue","darkred")}
}
\value{
ggplot volcano plot
}
\description{
Create an annotated volcano plot
}
