% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcano_plot_cuffdiff}
\alias{volcano_plot_cuffdiff}
\title{Create an annotated volcano plot from cuffdiff data}
\usage{
volcano_plot_cuffdiff(diffexp_data, group1, group2, ...)
}
\arguments{
\item{diffexp_data}{data frame of gene differential expression, must contain
column for gene IDs/symbols, log2 fold-changes and a p-value}

\item{group1}{character string, name of experimental group in the data.}

\item{group2}{character string, name of group in the data that will be used
as a baseline.}

\item{...}{other parameters to plot_volcano function}
}
\value{
ggplot volcano plot
}
\description{
Create an annotated volcano plot from cuffdiff data
}
