% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{ggsave_param_wrapper}
\alias{ggsave_param_wrapper}
\title{Wrapper around ggsave_param function}
\usage{
ggsave_param_wrapper(graph_type, output_directory, output_config_file, ...)
}
\arguments{
\item{graph_type}{character string. Type of graph to be saved, needs to be a
property in the config file}

\item{output_directory}{character string. Path to the output directory where
the plots will be saved}

\item{output_config_file}{character string. Path to json config file containing
graph dimensions for specified graph type}

\item{...}{other parameters passed to the ggsave_param function}
}
\description{
To make saving graphs easier, only graph_type property from the config file
needs to be supplied. Functions assumes there is a file called 'config.json'
in the current directory and a output_directory global variable that points
to the locations graphs will be saved in.
}
