% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Generic PCA plotting function}
\usage{
plot_pca(
  pca_data,
  percent_var_labs,
  group = "group",
  plot_center = TRUE,
  linetype = "solid",
  palette = NULL
)
}
\arguments{
\item{pca_data}{data frame of PCA data, with required PC1 and PC2 columns}

\item{percent_var_labs}{double vector of length 2, numeric values of
explained variance for the PCA plot axes}

\item{plot_center}{logical, whether to plot the center of group in the PCA
plot}

\item{linetype}{character string, a valid linetype for ggplot2. If the
plot_center is set to TRUE, this will determine the line type of drawn
connections between sample points to group center. default: solid
be}

\item{palette}{character vector of the same length as the number of groups in
the data. Custom color palette for groups. If NULL, viridis color palette
will be used. default: NULL}
}
\value{
ggplot PCA plot
}
\description{
Generic PCA plotting function
}
